requirements.txt
#crewai
#crewai[tools]
crewai==0.28.8
crewai-tools==0.1.6
langchain-openai==0.1.0
langchain-community==0.0.29
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2